--- src/com/massivecraft/creativegates/CreativeGates.java
@@ -1,103 +1,0 @@
-package com.massivecraft.creativegates;
-
-import java.util.EnumSet;
-import java.util.Set;
-
-import org.bukkit.Material;
-import org.bukkit.block.Block;
-
-import com.massivecraft.creativegates.entity.UConfColls;
-import com.massivecraft.creativegates.entity.UGateColls;
-import com.massivecraft.creativegates.index.IndexCombined;
-import com.massivecraft.mcore.Aspect;
-import com.massivecraft.mcore.AspectColl;
-import com.massivecraft.mcore.MPlugin;
-import com.massivecraft.mcore.Multiverse;
-
-public class CreativeGates extends MPlugin
-{
-	// -------------------------------------------- //
-	// CONSTANTS
-	// -------------------------------------------- //
-	
-	public final static Set<Material> VOID_MATERIALS = EnumSet.of(Material.AIR, Material.WATER, Material.STATIONARY_WATER, Material.LAVA, Material.STATIONARY_LAVA); 
-	
-	// -------------------------------------------- //
-	// INSTANCE & CONSTRUCT
-	// -------------------------------------------- //
-	
-	private static CreativeGates i;
-	public static CreativeGates get() { return i; }
-	public CreativeGates() { CreativeGates.i = this; }
-	
-	// -------------------------------------------- //
-	// FIELDS
-	// -------------------------------------------- //
-	
-	// Aspects
-	private Aspect aspect;
-	public Aspect getAspect() { return this.aspect; }
-	public Multiverse getMultiverse() { return this.getAspect().getMultiverse(); }
-	
-	// Index
-	private IndexCombined index = new IndexCombined();
-	public IndexCombined getIndex() { return this.index; };
-	
-	// Filling
-	private boolean filling = false;
-	public boolean isFilling() { return this.filling; }
-	public void setFilling(boolean filling) { this.filling = filling; }
-	
-	// -------------------------------------------- //
-	// OVERRIDE
-	// -------------------------------------------- //
-	
-	@Override
-	public void onEnable()
-	{
-		if ( ! preEnable()) return;
-		
-		// Load Server Config
-		ConfServer.get().load();
-		
-		// Initialize Aspects
-		this.aspect = AspectColl.get().get(Const.ASPECT_ID, true);
-		this.aspect.register();
-		this.aspect.setDesc(
-			"<i>What gates do exist.",
-			"<i>What the config options are set to."
-		);
-
-		// Initialize Database
-		this.getIndex().clear();
-		UConfColls.get().init();
-		UGateColls.get().init();
-		
-		// Setup Listeners
-		MainListener.get().activate();
-		
-		postEnable();
-	}
-	
-	@Override
-	public void onDisable()
-	{
-		this.getIndex().clear();
-		super.onDisable();
-	}
-	
-	// -------------------------------------------- //
-	// UTIL
-	// -------------------------------------------- //
-	
-	public static boolean isVoid(Material material)
-	{
-		return VOID_MATERIALS.contains(material);
-	}
-	
-	public static boolean isVoid(Block block)
-	{
-		return isVoid(block.getType());
-	}
-	
-}
