--- src/com/massivecraft/creativegates/Perm.java
@@ -1,47 +1,0 @@
-package com.massivecraft.creativegates;
-
-import org.bukkit.command.CommandSender;
-
-import com.massivecraft.mcore.util.PermUtil;
-
-public enum Perm
-{
-	// -------------------------------------------- //
-	// ENUM
-	// -------------------------------------------- //
-	
-	CREATE("create"),
-	USE("use"),
-	
-	// END OF LIST
-	;
-	
-	// -------------------------------------------- //
-	// FIELDS
-	// -------------------------------------------- //
-	
-	public final String node;
-	
-	// -------------------------------------------- //
-	// CONSTRUCT
-	// -------------------------------------------- //
-	
-	Perm(final String node)
-	{
-		this.node = "creativegates."+node;
-	}
-	
-	// -------------------------------------------- //
-	// HAS
-	// -------------------------------------------- //
-	
-	public boolean has(CommandSender sender, boolean informSenderIfNot)
-	{
-		return PermUtil.has(sender, this.node, informSenderIfNot);
-	}
-	
-	public boolean has(CommandSender sender)
-	{
-		return has(sender, false);
-	}
-}
