--- src/com/massivecraft/creativegates/index/IndexAbstract.java
@@ -1,6 +1,0 @@
-package com.massivecraft.creativegates.index;
-
-public abstract class IndexAbstract implements Index
-{
-
-}
