+++ src/main/java/com/massivecraft/creativegates/Const.java
@@ -1,0 +1,14 @@
+package com.massivecraft.creativegates;
+
+public class Const
+{
+	// MStore Collection Names
+	public static final String COLLECTION_BASENAME = "creativegates";
+	public static final String COLLECTION_BASENAME_ = COLLECTION_BASENAME+"_";
+	public static final String COLLECTION_BASENAME_UCONF = COLLECTION_BASENAME_+"uconf";
+	public static final String COLLECTION_BASENAME_UGATE = COLLECTION_BASENAME_+"ugate";
+	
+	// Aspect Ids
+	public static final String ASPECT_ID = "creativegates";
+	
+}
