+++ src/main/java/com/massivecraft/creativegates/index/IndexAbstract.java
@@ -1,0 +1,6 @@
+package com.massivecraft.creativegates.index;
+
+public abstract class IndexAbstract implements Index
+{
+
+}
